[trim_paired] opt to output singles to same file
Currently, sickle is hard to use in the middle of a pipeline. If one
wanted to pipe output to other commands, it would be most convenient to
allow downstream commands to remove single reads, and not split the
pipeline three ways at the sickle step. I.e., this patch allows one to:

    pairs join 1.fq 2.fq | sickle pe -c /dev/stdin -t sanger -M /dev/stdout |\
        seqqs -i -e -p post_sickle /dev/stdin |\
        pairs split -1 out1.fq -2 out2.fq -u singles.fq